新增优惠劵！！！

+++ mall-pojo/src/main/java/cn/dnaizn/mall/VO/SellerVO.java
@@ -68,7 +68,7 @@ public class SellerVO implements Serializable {
 
     private BigDecimal deliverFee;
 
-    private Integer deliverDiscount;
+    private BigDecimal deliverDiscount;
 
     private Object businessHours;
